Start splitting out some types
